15_Bigdata/bigdata-sub3-master/backend/requirements.txt
appnope==0.1.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
chardet==3.0.4
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
Django==2.2.4
django-appconf==1.0.3
django-bootstrap4==0.0.8
django-cors-headers==3.1.0
django-debug-toolbar==1.11
django-extensions==2.1.6
django-imagekit==4.0.2
django-pandas==0.6.1
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
entrypoints==0.3
Faker==1.0.4
idna==2.8
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
nbconvert==5.4.1
nbformat==4.4.0
nose==1.3.7
notebook==5.7.8
numpy==1.17.0
pandas==0.25.1
pandocfilters==1.4.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
pilkit==2.0
Pillow==6.0.0
prometheus-client==0.6.0
prompt-toolkit==2.0.9
ptyprocess==0.6.0
Pygments==2.3.1
PyJWT==1.7.1
pyparsing==2.4.0
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.2
pyzmq==18.0.1
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
tornado==6.0.2
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.2
widgetsnbextension==3.4.2
